feat: Replace legacy MainActivity with new main screen
