+++ koin/src/main/AndroidManifest.xml
@@ -101,18 +101,6 @@
             </intent-filter>
         </activity>
 
-        <activity
-            android:name=".ui.newmain.NewMainActivity"
-            android:exported="true"
-            android:label="@string/app_name"
-            android:theme="@style/KAPTheme.NoActionBar">
-            <intent-filter>
-                <action android:name="android.intent.action.VIEW"/>
-                <category android:name="android.intent.category.DEFAULT"/>
-                <data android:host="newmain" android:pathPrefix="/navigation" android:scheme="koin"/>
-            </intent-filter>
-        </activity>
-
         <meta-data
             android:name="com.naver.maps.map.NCP_KEY_ID"
             android:value="${naverMapKey}" />
