+++ koin/src/main/java/in/koreatech/koin/ui/main/fragment/NotificationFragment.kt
@@ -1,4 +1,4 @@
-package `in`.koreatech.koin.ui.newmain.fragment
+package `in`.koreatech.koin.ui.main.fragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -9,12 +9,11 @@ import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import dagger.hilt.android.AndroidEntryPoint
-import `in`.koreatech.koin.R
 import `in`.koreatech.koin.core.designsystem.theme.RebrandKoinTheme
-import `in`.koreatech.koin.feature.home.HomeScreen
+import `in`.koreatech.koin.feature.notification.NotificationScreen
 
 @AndroidEntryPoint
-class HomeFragment : Fragment() {
+class NotificationFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
@@ -24,10 +23,8 @@ class HomeFragment : Fragment() {
             setViewCompositionStrategy(ViewCompositionStrategy.DisposeOnViewTreeLifecycleDestroyed)
             setContent {
                 RebrandKoinTheme {
-                    HomeScreen(
-                        navigateToNotification = {
-                            findNavController().navigate(R.id.action_home_to_notification)
-                        }
+                    NotificationScreen(
+                        onBack = { findNavController().popBackStack() }
                     )
                 }
             }
